requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
gspread==5.11.2
oauth2client==4.1.3
python-dotenv==1.0.0